Move the error handlers related to `account-recovery` error actions to its own subdirectory to make the middleware part less clutter.
- `dispatchSuccess` and `dispatchError` are also moved to its own
  module, `utils.js`, to prevent cycle dependency.
